tanxiao

+++ src/edu/cqun/eshop/constant/OrderStatus.java
@@ -1,10 +1,10 @@
 package edu.cqun.eshop.constant;
 
 public class OrderStatus {
-	public final static  int TROLLEY=1;
-	public final static  int NON_PAY=2;
-	public final static  int PAID=3;
-	public final static  int DELIVERING=4;
-	public final static  int DONE=5;
-	public final static  int COMMENTED=6;
+	public final static  short TROLLEY=1;
+	public final static  short NON_PAY=2;
+	public final static  short PAID=3;
+	public final static  short DELIVERING=4;
+	public final static  short DONE=5;
+	public final static  short COMMENTED=6;
 }
